feat: refactor delegation cli for modularity
