+++ bolt-cli/src/types.rs
@@ -2,18 +2,6 @@ use alloy::signers::k256::sha2::{Digest, Sha256};
 use ethereum_consensus::crypto::{PublicKey as BlsPublicKey, Signature as BlsSignature};
 use serde::Serialize;
 
-#[derive(Debug, thiserror::Error)]
-pub enum KeystoreError {
-    #[error("failed to read keystore directory: {0}")]
-    ReadFromDirectory(#[from] std::io::Error),
-    #[error("Failed to read or decrypt keystore: {0:?}")]
-    Eth2Keystore(lighthouse_eth2_keystore::Error),
-    #[error("Failed to get public key from keypair: {0}")]
-    UnknownPublicKey(String),
-    #[error("Missing password for keypair")]
-    MissingPassword,
-}
-
 /// Event types that can be emitted by the validator pubkey to
 /// signal some action on the Bolt protocol.
 #[derive(Debug, Clone, Copy)]
